remove test for file to not exist on chmod 000
this is a behavior difference between macos and linux, I guess. on
linux, the file is still returned as existing, and on macos the file
isn't returned at all for some reason.

I would argue that this test is incorrect, and we should in fact admit
the file exists even if we can't read or write it.