nit: Derive Ord
The current implementation of Ord is actually
identical to what `derive(Ord, PartialOrd)` would do.
So IMO we may as well just derive it, it's less
code we need to read and worry about.

Obviously if you were aware of this and wanted
it to be written out manually that's fine, just
close this PR :)