perf: Remove unnecessary tokio::io::copy
When downloading a Ruby tarball, each chunk of the
tarball is written to a file. Currently on main,
this uses tokio::io::copy. But that's unnecessary
as there's already a `write_all` method on the
chunk.

Benchmark shows that removing the io::copy and
using the write_all is somewhat faster:

```
hyperfine -w1 \
'rv ruby install 3.4.1 -n' \
'./target/release/rv ruby install 3.4.1 -n'

./target/release/rv ruby install 3.4.1 -n ran
1.21 ± 0.30 times faster than rv ruby install 3.4.1 -n
```