Create tarball file *after* downloading tarball
Currently, the tarball file is created on disk,
then the tarball itself is downloaded, and written
to that file.

Problem is, if the download fails, the empty
file is left on disk. That means downloads get
skipped in the future, and the invalid tarball
causes errors when you try to extract it.

Solution: simply defer creating the file until
after we've downloaded the tarball that will be
written into it.

Fixes https://github.com/spinel-coop/rv/issues/86.