Prefer ChecksumAlgorithm over strings in more places
The internals of rv-gem-package use ChecksumAlgorithm, an enum, in
many places. But also, in a lot of places, it works with the stringified
representation of those algorithms instead.

Working with the enums would be more type-safe (because you don't have
to parse the enums out of the strings later) and also more efficient
(because the enums don't require any stack allocations).

Also slightly bumped test coverage from 76% to 77%.