Stream files through hasher instead of buffering entirely into RAM.
Previously, to check the checksums, I was reading the entire file
into memory and then hashing it. This obviously can take up a lot more
memory than just handling the file one chunk at a time (e.g. reading
each entry out of the data.tar.gz and unpacking it).

To get the best of both worlds, I introduced a hashing reader, which
computes the sha256 and sha512 hashes as we're reading from the
data source. Then we finalize it at the end.