Let `rv ruby install` without explicit or incomplete version request work
If no explicit ruby version request is given as an argument, it will
automatically install the pinned version in a `.ruby-version` file or
equivalent if any, or latest stable version otherwise.

Similarly, if an "incomplete" ruby version request is given, like `rv
ruby install 4`, it will install the latest stable version matching the
request.