Let `rv ruby list` skip listing ruby 3.5.0-preview1
Subsequent prereleases and the final Ruby 3.5 release were skipped and
release numbering went straight to 4.0. It's uninteresting for users to
always see this release, so let's keep only the latest stable release
for each series.

This preview version remains installable and we may add an `--all` flag
to show it, but by default it should be hidden.